refactor : SectionUpdateType의 데이터를 Dto로 이동
enum class의 동시성 문제를 해결하기 위함
